pyproject.toml
[project]
name = "fastbin-restart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "authlib>=1.6.5",
    "celery>=5.5.3",
    "fastapi[standard]>=0.122.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "prometheus-client>=0.23.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=7.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "anyio>=4.11.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]


